refactor: move `useSortAndFilterPathoscope` to `hooks.ts`
