+++ src/js/analyses/components/Pathoscope/PathoscopeDetail.tsx
@@ -1,11 +1,19 @@
+import { FormattedPathoscopeHit } from "@/analyses/types";
 import { useUrlSearchParams } from "@utils/hooks";
 import { filter, map } from "lodash-es";
 import React from "react";
 import { ScrollSync } from "react-scroll-sync";
 import { PathoscopeIsolate } from "./Isolate";
 
+type PathoscopeDetailProps = {
+    /** Complete information for a pathoscope hit */
+    hit: FormattedPathoscopeHit;
+    /** The total number of reads mapped to any OTU during the analysis*/
+    mappedCount: number;
+};
+
 /** Detailed coverage for a single OTU hits from pathoscope analysis*/
-export const PathoscopeDetail = ({ hit, mappedCount }) => {
+export function PathoscopeDetail({ hit, mappedCount }: PathoscopeDetailProps) {
     const [filterIsolates] = useUrlSearchParams<boolean>("filterIsolates");
     const [showReads] = useUrlSearchParams<boolean>("reads");
     const { isolates, pi } = hit;
@@ -27,4 +35,4 @@ export const PathoscopeDetail = ({ hit, mappedCount }) => {
             </ScrollSync>
         </div>
     );
-};
+}
